+++ packages/orq-rc/src/orq_ai_sdk/datasets.py
@@ -12,7 +12,6 @@ class Datasets(BaseSDK):
     def list(
         self,
         *,
-        sort: Optional[models.ListDatasetsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -25,7 +24,6 @@ class Datasets(BaseSDK):
 
         Retrieves a paginated list of datasets for the current workspace. Results can be paginated using cursor-based pagination.
 
-        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -48,7 +46,6 @@ class Datasets(BaseSDK):
             base_url = self._get_url(base_url, url_variables)
 
         request = models.ListDatasetsRequest(
-            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,
@@ -119,7 +116,6 @@ class Datasets(BaseSDK):
     async def list_async(
         self,
         *,
-        sort: Optional[models.ListDatasetsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -132,7 +128,6 @@ class Datasets(BaseSDK):
 
         Retrieves a paginated list of datasets for the current workspace. Results can be paginated using cursor-based pagination.
 
-        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -155,7 +150,6 @@ class Datasets(BaseSDK):
             base_url = self._get_url(base_url, url_variables)
 
         request = models.ListDatasetsRequest(
-            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,
@@ -1067,7 +1061,6 @@ class Datasets(BaseSDK):
         self,
         *,
         dataset_id: str,
-        sort: Optional[models.ListDatasetDatapointsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -1081,7 +1074,6 @@ class Datasets(BaseSDK):
         Retrieves a paginated list of datapoints from a specific dataset.
 
         :param dataset_id:
-        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -1105,7 +1097,6 @@ class Datasets(BaseSDK):
 
         request = models.ListDatasetDatapointsRequest(
             dataset_id=dataset_id,
-            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,
@@ -1177,7 +1168,6 @@ class Datasets(BaseSDK):
         self,
         *,
         dataset_id: str,
-        sort: Optional[models.ListDatasetDatapointsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -1191,7 +1181,6 @@ class Datasets(BaseSDK):
         Retrieves a paginated list of datapoints from a specific dataset.
 
         :param dataset_id:
-        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -1215,7 +1204,6 @@ class Datasets(BaseSDK):
 
         request = models.ListDatasetDatapointsRequest(
             dataset_id=dataset_id,
-            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,