+++ packages/orq-rc/src/orq_ai_sdk/models/bulkcreatedatapointsop.py
@@ -387,5 +387,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-21T10:08:06.343Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-25T10:15:15.421Z")
     r"""The date and time the resource was last updated"""