+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -10,13 +10,7 @@ from typing import List, Literal, Optional
 from typing_extensions import Annotated, NotRequired, TypedDict
 
 
-QueryParamSort = Literal["asc", "desc"]
-r"""List sorting preference."""
-
-
 class FileListRequestTypedDict(TypedDict):
-    sort: NotRequired[QueryParamSort]
-    r"""List sorting preference."""
     limit: NotRequired[float]
     r"""A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10"""
     starting_after: NotRequired[str]
@@ -26,12 +20,6 @@ class FileListRequestTypedDict(TypedDict):
 
 
 class FileListRequest(BaseModel):
-    sort: Annotated[
-        Optional[QueryParamSort],
-        FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
-    ] = "asc"
-    r"""List sorting preference."""
-
     limit: Annotated[
         Optional[float],
         FieldMetadata(query=QueryParamMetadata(style="form", explode=True)),
@@ -87,7 +75,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-21T10:08:07.574Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-25T10:15:17.066Z")
     r"""The date and time the resource was created"""